Add function to create project list
